requirements.txt
#Pager duty libs
authlib
boto3
dateparser
python-dotenv
gspread
python-dateutil
pytz
pyopenssl
PyAthena
requests
slackclient
